refactor: move flattenPaginatedPages to modules/queries/
